arquitect-events/requirements.txt
stomp.py
jsonschema
pytest
pytest-mock
pytest-cov